refactored code, added license, added onError handler
