+++ framework/src/com/phonegap/websocket/WebSocketFactory.java
@@ -1,15 +1,42 @@
+/*
+ * Copyright (c) 2010 Animesh Kumar  (https://github.com/anismiles)
+ *  
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
+ *  
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *  
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ *  
+ */
 package com.phonegap.websocket;
 
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Random;
 
-
 import android.webkit.WebView;
 
 /**
- * @author http://github.com/anismiles
- *
+ * The <tt>WebSocketFactory</tt> is like a helper class to instantiate new
+ * WebSocket instaces especially from Javascript side. It expects a valid
+ * "ws://" URI.
+ * 
+ * @author Animesh Kumar
  */
 public class WebSocketFactory {
 
@@ -18,25 +45,31 @@ public class WebSocketFactory {
 
 	/**
 	 * Instantiates a new web socket factory.
-	 *
-	 * @param appView the app view
+	 * 
+	 * @param appView
+	 *            the app view
 	 */
 	public WebSocketFactory(WebView appView) {
 		this.appView = appView;
 	}
 
-	/**
-	 * Gets the web socket.
-	 *
-	 * @param url the url
-	 * @return the web socket
-	 * @throws URISyntaxException the uRI syntax exception
-	 */
 	public WebSocket getInstance(String url) throws URISyntaxException {
-		// random id
-		String id = "WebSocket." + new Random().nextInt(100);
-		WebSocket socket =  new WebSocket(appView, new URI(url), Protocol.Draft.DRAFT75, id);
+		return getInstance(url, WebSocket.Draft.DRAFT75);
+	}
+
+	public WebSocket getInstance(String url, WebSocket.Draft draft) throws URISyntaxException {
+		WebSocket socket = new WebSocket(appView, new URI(url), draft, getRandonUniqueId());
 		socket.connect();
 		return socket;
 	}
+
+	/**
+	 * Generates random unique ids for WebSocket instances
+	 * 
+	 * @return String
+	 */
+	private String getRandonUniqueId() {
+		return "WEBSOCKET." + new Random().nextInt(100);
+	}
+
 }
